+++ test/td_dq_web/controllers/implementation_controller_test.exs
@@ -3602,15 +3602,44 @@ defmodule TdDqWeb.ImplementationControllerTest do
 
   describe "csv" do
     setup do
+      details = %{
+        "Query" => "Rk9P",
+        "baz_title" => "baz"
+      }
+
+      details2 = %{
+        "foo_title" => %{"x" => "foo"},
+        "baz_title" => "bazz",
+        "jaz_title" => "jaz"
+      }
+
       result =
-        build(:rule_result, records: 3245, result_type: "percentage", errors: 123, result: 0)
+        build(:rule_result,
+          records: 3245,
+          result_type: "percentage",
+          errors: 123,
+          result: 0,
+          details: details
+        )
+
+      result2 =
+        build(:rule_result,
+          records: 3245,
+          result_type: "percentage",
+          errors: 123,
+          result: 0,
+          details: details2
+        )
 
       implementations = [
         insert(:implementation,
           results: [result],
           df_content: %{"some_first_field" => "some_first_value"}
         ),
-        insert(:implementation, df_content: %{"some_second_field" => "some_value"}),
+        insert(:implementation,
+          results: [result2],
+          df_content: %{"some_second_field" => "some_value"}
+        ),
         insert(:implementation, df_content: %{"some_second_field" => "some_second_value"})
       ]
 
@@ -3680,11 +3709,90 @@ defmodule TdDqWeb.ImplementationControllerTest do
 
       for regex <- [
             # credo:disable-for-lines:5 Credo.Check.Readability.MaxLineLength
+            "implementation_key;implementation_type;executable;rule;rule_template;implementation_template;goal;minimum;business_concept;last_execution_at;records;errors;result;execution;inserted_at;result_details_Query;result_details_baz_title;result_details_foo_title;result_details_jaz_title;dataset_external_id_1;validation_field_1\r",
+            ~r/#{key_0};default;[\w+.]+;#{name_0};;;\d*\.?\d*;\d*\.?\d*;;;;;;;#{ts_0};;;;;;\r/,
+            ~r/#{key_1};default;[\w+.]+;#{name_1};;;\d*\.?\d*;\d*\.?\d*;;[[:ascii:]]+;#{records_1};#{errors_1};\d*\.?\d*;[\w+.]+;#{ts_1};FOO;baz;;;;\r/,
+            ~r/#{key_2};default;[\w+.]+;#{name_2};;;\d*\.?\d*;\d*\.?\d*;;[[:ascii:]]+;#{records_1};#{errors_1};\d*\.?\d*;[\w+.]+;#{ts_2};;bazz;\"{\"\"x\"\":\"\"foo\"\"}\";jaz;;\r/,
+            ~r/#{key_3};default;[\w+.]+;#{name_3};;;\d*\.?\d*;\d*\.?\d*;;;;;;;#{ts_3};;;;;;\r/
+          ] do
+        assert body =~ regex
+      end
+    end
+
+    @tag authentication: [role: "admin"]
+    test "download implementations with result details only for admin", %{
+      conn: conn,
+      implementations: implementations
+    } do
+      expect(ElasticsearchMock, :request, fn _, :post, "/implementations/_search", _, _ ->
+        SearchHelpers.hits_response(implementations)
+      end)
+
+      [
+        %{
+          implementation_key: key_1,
+          rule: %{name: name_1},
+          results: [
+            %{
+              records: records_1,
+              errors: errors_1,
+              details: %{"Query" => query_base64, "baz_title" => detail_field1}
+            }
+          ]
+        },
+        %{
+          implementation_key: key_2,
+          rule: %{name: name_2},
+          results: [%{details: %{"baz_title" => baz_title, "jaz_title" => jaz_title}}]
+        },
+        %{implementation_key: key_3, rule: %{name: name_3}}
+      ] = implementations
+
+      {:ok, query} = Base.decode64(query_base64)
+
+      assert %{resp_body: body} = post(conn, Routes.implementation_path(conn, :csv, %{}))
+
+      for regex <- [
+            # credo:disable-for-lines:5 Credo.Check.Readability.MaxLineLength
+            "implementation_key;implementation_type;executable;rule;rule_template;implementation_template;goal;minimum;business_concept;last_execution_at;records;errors;result;execution;inserted_at;result_details_Query;result_details_baz_title;result_details_foo_title;result_details_jaz_title;dataset_external_id_1;validation_field_1\r",
+            ~r/#{key_1};default;[\w+.]+;#{name_1};;;\d*\.?\d*;\d*\.?\d*;;[[:ascii:]]+;#{records_1};#{errors_1};\d*\.?\d*;[\w+.]+;[[:ascii:]]+;#{query};#{detail_field1};;;;\r/,
+            ~r/#{key_2};default;[\w+.]+;#{name_2};;;\d*\.?\d*;\d*\.?\d*;;[[:ascii:]]+;#{records_1};#{errors_1};\d*\.?\d*;[\w+.]+;[[:ascii:]]+;;#{baz_title};\"{\"\"x\"\":\"\"foo\"\"}\";#{jaz_title};;\r/,
+            ~r/#{key_3};default;[\w+.]+;#{name_3};;;\d*\.?\d*;\d*\.?\d*;;;;;;;[[:ascii:]]+;;;;;;\r/
+          ] do
+        assert body =~ regex
+      end
+    end
+
+    @tag authentication: [
+           role: "non-admin",
+           permissions: @rule_implementation_permissions ++ [:manage_segments]
+         ]
+    test "download implementations without result details for non-admin", %{
+      conn: conn,
+      implementations: implementations
+    } do
+      expect(ElasticsearchMock, :request, fn _, :post, "/implementations/_search", _, _ ->
+        SearchHelpers.hits_response(implementations)
+      end)
+
+      [
+        %{
+          implementation_key: key_1,
+          rule: %{name: name_1},
+          results: [%{records: records_1, errors: errors_1}]
+        },
+        %{implementation_key: key_2, rule: %{name: name_2}},
+        %{implementation_key: key_3, rule: %{name: name_3}}
+      ] = implementations
+
+      assert %{resp_body: body} = post(conn, Routes.implementation_path(conn, :csv, %{}))
+
+      for regex <- [
+            # credo:disable-for-lines:5 Credo.Check.Readability.MaxLineLength
             "implementation_key;implementation_type;executable;rule;rule_template;implementation_template;goal;minimum;business_concept;last_execution_at;records;errors;result;execution;inserted_at;dataset_external_id_1;validation_field_1\r",
-            ~r/#{key_0};default;[\w+.]+;#{name_0};;;\d*\.?\d*;\d*\.?\d*;;;;;;;#{ts_0};;\r/,
-            ~r/#{key_1};default;[\w+.]+;#{name_1};;;\d*\.?\d*;\d*\.?\d*;;[[:ascii:]]+;#{records_1};#{errors_1};\d*\.?\d*;[\w+.]+;#{ts_1};;\r/,
-            ~r/#{key_2};default;[\w+.]+;#{name_2};;;\d*\.?\d*;\d*\.?\d*;;;;;;;#{ts_2};;\r/,
-            ~r/#{key_3};default;[\w+.]+;#{name_3};;;\d*\.?\d*;\d*\.?\d*;;;;;;;#{ts_3};;\r/
+            ~r/#{key_1};default;[\w+.]+;#{name_1};;;\d*\.?\d*;\d*\.?\d*;;[[:ascii:]]+;#{records_1};#{errors_1};\d*\.?\d*;[\w+.]+;[[:ascii:]]+;;\r/,
+            ~r/#{key_2};default;[\w+.]+;#{name_2};;;\d*\.?\d*;\d*\.?\d*;;[[:ascii:]]+;#{records_1};#{errors_1};\d*\.?\d*;[\w+.]+;[[:ascii:]]+;;\r/,
+            ~r/#{key_3};default;[\w+.]+;#{name_3};;;\d*\.?\d*;\d*\.?\d*;;;;;;;[[:ascii:]]+;;\r/
           ] do
         assert body =~ regex
       end